Fixed merge from branch 'jetty-9.3.x'.

+++ jetty-http2/http2-common/src/main/java/org/eclipse/jetty/http2/HTTP2Session.java
@@ -1114,7 +1114,7 @@ public abstract class HTTP2Session extends ContainerLifeCycle implements ISessio
 
     private class ControlEntry extends HTTP2Flusher.Entry
     {
-        private long bytes;
+        private int bytes;
 
         private ControlEntry(Frame frame, IStream stream, Callback callback)
         {
